requirements.txt
matplotlib
numpy
attrs
torchcde
scikit-learn